Initial Analysis of the Data
General Observations
The Data doesn't have a specified index; rather a sequence number.

Out of a total 5727 cases in Gibraltar, 4907 were hospitalized, 4670 recovered, and 97 died.

The number of hospitalization in Gibraltar is extremely high relative to the number of total cases.

The DataFrame has 8 missing values: 2 in the deaths column, 2 in the cases column, 2 in the recovered column, and 2 in the hospitalised column. However, as determined below, the data belongs to 2 instances on 2020-09-21 and 2020-09-22 in Bermuda. The reason for the missing data is likely due to the Tropical Storm Teddy, which hit Bermuda on the 21st of September. Teddy's centre was located offshore approximately 310 km south-east of Bermuda main island, with maximum sustained wind of 165 km/h.

More people have taken the first dose of vaccination than the second.